refactor: hooks hygiene — remove dead scaffolds, rename connection view
- Delete unused _protocolCall identity hooks (useSet*, useReboot, useGps,
  etc.) superseded by runtime panel actions and useSendMessage.
- Rename useLegacyConnectionView to useConnectionView (store-only reads).
- Move meshcoreHookTypes to lib/meshcore/; update component and runtime imports.
- Merge useLegacyConnectionActions.test into useProtocolConnection.test.
- Document hooks vs runtime vs lib boundaries in AGENTS.md and ARCHITECTURE.md.

Completes post-#375 hooks folder cleanup: no behavior change beyond removing
unreachable scaffold code; active API is runtime + useProtocolFacade.
