+++ src/renderer/components/MeshcoreTelemetryPrivacySection.tsx
@@ -2,7 +2,7 @@
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import type { MeshCoreContactRaw, MeshCoreSelfInfo } from '../hooks/meshcore/meshcoreHookTypes';
+import type { MeshCoreContactRaw, MeshCoreSelfInfo } from '../lib/meshcore/meshcoreHookTypes';
 import {
   countMeshcoreContactsWithFlagMask,
   MESHCORE_CONTACT_FLAG_TELEM_BASE,
