+++ src/renderer/hooks/useProtocolConnection.test.ts
@@ -7,8 +7,15 @@ import {
   type MeshtasticSessionApi,
   registerMeshtasticSession,
 } from '../lib/sessions/meshtasticSession';
-import { addIdentity } from '../stores/identityStore';
-import { useProtocolConnect, useProtocolDisconnect } from './useProtocolConnection';
+import { setConnection, useConnectionStore } from '../stores/connectionStore';
+import { addIdentity, useIdentityStore } from '../stores/identityStore';
+import {
+  useProtocolConnect,
+  useProtocolConnectionActions,
+  useProtocolDisconnect,
+} from './useProtocolConnection';
+
+const IDENTITY_ACTIONS = 'id-conn-actions-mt';
 
 const mockDriverConnect = vi.fn().mockResolvedValue('id-meshtastic-driver');
 const mockDriverDisconnect = vi.fn().mockResolvedValue(undefined);
@@ -95,5 +102,61 @@ describe('useProtocolDisconnect (driver-first)', () => {
 
     expect(meshtastic.finalizeDriverDisconnect).toHaveBeenCalled();
     expect(mockDriverDisconnect).toHaveBeenCalledWith('id-meshtastic-test');
+  });
+});
+
+describe('useProtocolConnectionActions', () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+    registerMeshtasticSession(null);
+    registerMeshcoreSession(null);
+    useConnectionStore.setState({ connections: {} });
+    useIdentityStore.setState({ identities: {}, activeIdentityId: null });
+    addIdentity({
+      id: IDENTITY_ACTIONS,
+      protocol: meshtasticProtocol,
+      signature: 'sig-conn',
+      transports: [],
+      createdAt: Date.now(),
+      lastSeenAt: Date.now(),
+    });
+  });
+
+  it('driver-first connect prepares and attaches Meshtastic session', async () => {
+    const meshtastic = createMeshtasticSessionStub();
+    registerMeshtasticSession(meshtastic);
+
+    const { result } = renderHook(() => useProtocolConnectionActions('meshtastic'));
+
+    await result.current.connect('serial', undefined, undefined);
+
+    expect(meshtastic.prepareRfConnect).toHaveBeenCalledWith('serial', undefined, undefined);
+    expect(meshtastic.attachRfSession).toHaveBeenCalledWith('id-meshtastic-driver', 'serial');
+  });
+
+  it('maps http to tcp for meshcore driver-first connect', async () => {
+    const meshcore = createMeshcoreSessionStub();
+    registerMeshcoreSession(meshcore);
+
+    const { result } = renderHook(() => useProtocolConnectionActions('meshcore'));
+
+    await result.current.connect('http', '192.168.1.1', undefined);
+
+    expect(meshcore.prepareRfConnect).toHaveBeenCalledWith('tcp');
+    expect(meshcore.attachRfSession).toHaveBeenCalledWith('id-meshtastic-driver', 'tcp');
+  });
+
+  it('exposes state from the connection store for the protocol identity', () => {
+    setConnection(IDENTITY_ACTIONS, {
+      status: 'configured',
+      connectionType: 'ble',
+      myNodeNum: 42,
+      mqttStatus: 'disconnected',
+    });
+
+    const { result } = renderHook(() => useProtocolConnectionActions('meshtastic'));
+
+    expect(result.current.state.myNodeNum).toBe(42);
+    expect(result.current.state.status).toBe('configured');
   });
 });
